add test; reorganize
