+++ msmexplorer/plots/__init__.py
@@ -1,4 +1,4 @@
 from .msm import *
-from .timescales import *
 from .tpt import *
 from .histogram import *
+from .cluster import *
