+++ msmexplorer/tests/test_msm_plot.py
@@ -3,7 +3,7 @@ from msmbuilder.msm import MarkovStateModel, BayesianMarkovStateModel
 from matplotlib.axes import Subplot
 from seaborn.apionly import JointGrid
 
-from ..plots import plot_pop_resids, plot_msm_network
+from ..plots import plot_pop_resids, plot_msm_network, plot_timescales
 
 data = np.random.randint(low=0, high=10, size=100000)
 msm = MarkovStateModel()
@@ -20,5 +20,17 @@ def test_plot_pop_resids():
 
 def test_plot_msm_network():
     ax = plot_msm_network(msm)
+
+    assert isinstance(ax, Subplot)
+
+
+def test_plot_timescales_msm():
+    ax = plot_timescales(msm)
+
+    assert isinstance(ax, Subplot)
+
+
+def test_plot_timescales_bmsm():
+    ax = plot_timescales(bmsm)
 
     assert isinstance(ax, Subplot)
